reorganize models refactor methods seperate datamanager uodate models fix club loading establish loading pattern make snapshot listener for comments to update automatically
