+++ MovieClub/Controllers/MovieManager.swift
@@ -1,0 +1,34 @@
+//
+//  MovieManager.swift
+//  MovieClub
+//
+//  Created by Marcus Lair on 9/24/24.
+//
+
+import Foundation
+
+extension DataManager {
+    
+    // MARK: - Fetch API Movie
+    
+    func fetchAPIMovie(title: String) async throws -> APIMovie {
+        let formattedTitle = title.replacingOccurrences(of: " ", with: "+")
+        let urlString = "https://omdbapi.com/?t=\(formattedTitle)&apikey=ab92d369"
+        guard let url = URL(string: urlString) else {
+            print("Invalid URL: \(urlString)")
+            throw URLError(.badURL)
+        }
+        let (data, response) = try await URLSession.shared.data(from: url)
+        guard let httpResponse = response as? HTTPURLResponse, httpResponse.statusCode == 200 else {
+            print("Bad server response: \(response)")
+            throw URLError(.badServerResponse)
+        }
+        let decoder = JSONDecoder()
+        do {
+            return try decoder.decode(APIMovie.self, from: data)
+        } catch {
+            print("Failed to decode API response: \(error)")
+            throw URLError(.cannotParseResponse)
+        }
+    }
+}
