+++ MovieClub/Views/Clubs/ComingSoon/ComingSoonRowView.swift
